// acsstats/AcsStatsWeb/ClientApp/src/app/modules/shared/models/victoryType.ts
export enum VictoryType {
  Awarded = "Awarded",
  Drawn = "Drawn",
  Runs = "Runs",
  Wickets = "Wickets",
  Innings = "Innings",
  Tied = "Tied",
  Abandoned = "Abandoned",
  NoResult = "NoResult",
  RunRate = "RunRate",
  LosingFewerWickets = "LosingFewerWickets",
  FasterScoringRate = "FasterScoringRate",
  Unknown = "Unknown"
}